requirements.txt
asgiref==3.4.1
certifi==2022.6.15
charset-normalizer==2.0.12
Django==3.2.14
djangorestframework==3.13.1
docker-py==1.10.6
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
gunicorn==20.1.0
idna==3.3
psycopg2-binary==2.9.3
pyrankvote==2.0.5
pytz==2022.1
PyYAML==6.0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tabulate==0.8.10
typing-extensions==4.1.1
urllib3==1.26.11
websocket-client==1.3.1
